+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## MultiQC v1.8dev
 
 #### New Modules:
+* [**fgbio**](http://fulcrumgenomics.github.io/fgbio/)
+    * Process family size count hist data from GroupReadsByUmi
 * [**biobambam2**](https://github.com/gt1/biobambam2)
     * Added submodule for `bamsormadup` tool
     * Totally cheating - it uses Picard MarkDuplicates but with a custom search pattern and naming